searxng-mcp-crawl/requirements.txt
starlette>=0.27.0
uvicorn>=0.23.0
httpx>=0.24.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
html5lib>=1.1
markdown>=3.5
